+++ src/control/e2e.spec/select-multiple.spec.ts
@@ -8,6 +8,7 @@ import { Observable } from 'rxjs/Observable';
 import { Subject } from 'rxjs/Subject';
 
 import { MarkAsDirtyAction, SetValueAction } from '../../actions';
+import { box, Boxed } from '../../boxing';
 import { NgrxValueConverters } from '../../control/value-converter';
 import { NgrxFormsModule } from '../../module';
 import { createFormControlState, FormControlState } from '../../state';
@@ -73,6 +74,82 @@ describe(SelectMultipleComponent.name, () => {
     actions$.first().subscribe(a => {
       expect(a.type).toBe(SetValueAction.TYPE);
       expect((a as SetValueAction<string>).value).toBe(JSON.stringify(SELECT_OPTIONS));
+      done();
+    });
+
+    option1.selected = true;
+    element.dispatchEvent(new Event('change'));
+  });
+
+  it(`should trigger a ${MarkAsDirtyAction.name} when an option is selected`, done => {
+    actions$.skip(1).first().subscribe(a => {
+      expect(a.type).toBe(MarkAsDirtyAction.TYPE);
+      done();
+    });
+
+    option1.selected = true;
+    element.dispatchEvent(new Event('change'));
+  });
+});
+
+@Component({
+  // tslint:disable-next-line:component-selector
+  selector: 'select-multiple-test',
+  template: `
+  <select multiple [ngrxFormControlState]="state">
+    <option *ngFor="let o of options" [value]="o">{{ o }}</option>
+  </select>
+  `,
+})
+export class SelectMultipleWithoutConverterComponent {
+  @Input() state: FormControlState<Boxed<string[]>>;
+  options = SELECT_OPTIONS;
+}
+
+describe(SelectMultipleWithoutConverterComponent.name, () => {
+  let component: SelectMultipleWithoutConverterComponent;
+  let fixture: ComponentFixture<SelectMultipleWithoutConverterComponent>;
+  let actionsSubject: ActionsSubject;
+  let actions$: Observable<Action>;
+  let element: HTMLSelectElement;
+  let option1: HTMLOptionElement;
+  let option2: HTMLOptionElement;
+  const FORM_CONTROL_ID = 'test ID';
+  const INITIAL_FORM_CONTROL_VALUE = box([SELECT_OPTIONS[1]]);
+  const INITIAL_STATE = createFormControlState(FORM_CONTROL_ID, INITIAL_FORM_CONTROL_VALUE);
+
+  beforeEach(() => {
+    actionsSubject = new Subject<Action>() as ActionsSubject;
+    actions$ = actionsSubject as Observable<Action>; // cast required due to mismatch of lift() function signature
+  });
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      imports: [NgrxFormsModule],
+      declarations: [SelectMultipleWithoutConverterComponent],
+      providers: [{ provide: ActionsSubject, useValue: actionsSubject }],
+    }).compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(SelectMultipleWithoutConverterComponent);
+    component = fixture.componentInstance;
+    component.state = INITIAL_STATE;
+    fixture.detectChanges();
+    const nativeElement = fixture.nativeElement as HTMLElement;
+    element = nativeElement.querySelector('select')!;
+    option1 = nativeElement.querySelectorAll('option')[0];
+    option2 = nativeElement.querySelectorAll('option')[1];
+  });
+
+  it('should select the correct option initially', () => {
+    expect(option2.selected).toBe(true);
+  });
+
+  it('should trigger a SetValueAction with the selected value when an option is selected', done => {
+    actions$.first().subscribe(a => {
+      expect(a.type).toBe(SetValueAction.TYPE);
+      expect((a as SetValueAction<Boxed<string[]>>).value).toEqual(box(SELECT_OPTIONS));
       done();
     });